+++ test/crockery/hiccup_test.clj
@@ -23,3 +23,22 @@
                          :align :center
                          :title-align :right}
                         :created] data)))))
+(deftest table-no-titles-test
+  (let [data [{:company "EvilCo" :created #inst "2019-03-01"}
+              {:company "Silly, Inc." :created #inst "2018-12-31"}]]
+    (is (=
+         [:table
+          [:tbody
+           (list
+            [:tr
+             [:td {:align "center"} "EvilCo"]
+             [:td {:align "left"} "2019-03-01T00:00:00Z"]]
+            [:tr
+             [:td {:align "center"} "Silly, Inc."]
+             [:td {:align "left"} "2018-12-31T00:00:00Z"]])]]
+         (hiccup/table {:titles? false}
+                       [{:name :company
+                         :align :center
+                         :title-align :right}
+                        :created]
+                       data)))))